fix region toggling
add `/regions` route with an action. hitting the
`/regions` endpoint with a POST will set the
"regions" cookie to only include regions that were
marked as being "on" in the form data provided.

wrap RegionToggle in a Form element and make it
submit a POST to the `/regions` route every time one
of the checkboxes it contains is clicked.