fix region toggling
add `/regions` route with an action. hitting the
`/regions` endpoint with a POST will set the
"regions" query parameter to only include
regions that were marked as being "on" in the
form data provided.

use refs to grab the values of every single input
in the region toggle on change, then submit to the
`/regions` endpoint.